Ensure that layer3 packets are displayed as well
When using e.g. WireGuard (a VPN which completely acts on layer3), no
packages will be matched as it's attempted to parse those as
ethernet (=layer2) packets. This is a problem as all layer3-packets fail
to get parsed properly (due to different offsets in the packet, wrong
protocols will be determined for instance).

This change inherits the basic idea from
`<libpnet/examples/packetdump.rs>` to check if it's possible to parse
version info using the IpPacket-parsers and if that fails, the sniffer
will fall-back to the ethernet-based approach.