Address review feedback
Now selectively printing out the duration of the capture based on
calculating the exact length of the strings and the width of the
terminal window. This has simplified this code substantially as well, no
longer having to split the header into parts (currently we are printing
to the same rect, but left and right aligned for the respective parts of
the header).